backend/requirements.txt
fastapi==0.104.1
pydantic==2.4.2
pydantic-settings==2.0.3
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Use a version with prebuilt wheels for Python 3.12 on Windows
aiohttp==3.9.5
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.1
websockets==12.0
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
pyOpenSSL==23.2.0
dnspython==2.4.2
lxml==4.9.3
psutil==5.9.8
python-owasp-zap-v2.4==0.0.20
typing-extensions==4.8.0
pandas==2.1.3
# Python 3.12 requires numpy >=1.26
numpy==1.26.4
PyJWT==2.8.0
builtwith>=1.3.4
requests>=2.31.0
vulture
watchdog
yara-python
zstandard
python-Wappalyzer
setuptools<81
# Additional dependencies for enhanced scanning capabilities
json5
python-whois
tldextract
